@@ -3,7 +3,7 @@ name = "zeroband"
 version = "0.1.0"
 description = "ZeroBand is a production ready codebase for decentralized training of LLM"
 readme = "README.md"
-requires-python = ">=3.10, <3.11"
+requires-python = "==3.11.*"
 dependencies = [
     "torch",
     "numpy",
@@ -27,11 +27,13 @@ dependencies = [
     "shardcast >= 0.3.1",
     "asyncio>=3.4.3",
     "aiohttp>=3.10.5",
-    "pyext>=0.7",
+    "pyext @ git+https://github.com/justusmattern27/PyExt.git",
     "prime-iroh>=0.2.1",
     "setuptools",
     "fastapi",
     "requests",
+    "reasoning-gym @ git+https://github.com/open-thought/reasoning-gym.git",
+    "tomli>=2.2.1",
 ] 
 
 [project.optional-dependencies]
@@ -48,8 +50,16 @@ allow-direct-references = true # allow direct references to git repos in depende
 line-length = 140
 
 [tool.ruff.lint]
+select = ["F", "I"]
 ignore = ["F722", "F821"]
 
 [tool.uv]
 dev-dependencies = ["ruff>=0.5.0", "pre-commit>=3.0.0","pytest>=7.0.0", "faker"]
 no-build-isolation-package = ["flash-attn"]
+
+[tool.pytest.ini_options]
+addopts = "--strict-markers"
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "gpu: marks tests as gpu (deselect with '-m \"not gpu\"')",
+]
