@@ -1,115 +1,65 @@
-import requests
-import os
-from pathlib import Path
-import uuid
-import numpy as np
-import torch
-from vllm import LLM, SamplingParams
-from pydantic_config import parse_argv
-import time
-from toploc.utils import sha256sum
-from safetensors import safe_open
-import torch.distributed as dist
 import json
+import multiprocessing as mp
+import os
+import shutil
+import time
+import uuid
+from pathlib import Path
 
-# from vllm.model_executor.model_loader
-from vllm.model_executor.model_loader.loader import _process_weights_after_loading
+# Import environment before any other imports
+# ruff: noqa: I001
+from zeroband.inference import envs
 
+import numpy as np
+import pyarrow.parquet as pq
+import requests
+import torch
+import torch.distributed as dist
+from datasets import load_dataset
+from pydantic_config import parse_argv
+from toploc.utils import sha256sum
+from vllm import LLM, SamplingParams
 
 from zeroband.inference.config import Config
-from zeroband.utils.logger import get_logger
-
-from zeroband.inference.toploc import setup_toploc_cache
+from zeroband.inference.parquet import get_parquet_table
 from zeroband.inference.pipeline import setup_pipeline
 from zeroband.inference.rewards import compute_vllm_rewards
 from zeroband.inference.parquet import get_parquet_table
-
+from zeroband.inference.toploc import setup_toploc_cache
+from zeroband.inference.utils import fake_chat_template, filter_data_by_prompt_length, generate_target_length_prompts, reload_model_weights
 
 from datasets import load_dataset
 import pyarrow.parquet as pq
 import multiprocessing as mp
 
 from zeroband.training.mp import EnvWrapper
+from zeroband.utils.logger import get_logger
 from zeroband.utils.metrics import PrimeMetric
-
-from zeroband.inference import envs
 
 # Global logger
 logger = get_logger("INFER")
 
 
-def fake_chat_template(messages):
-    formatted_prompts = []
-
-    for conversation in messages:
-        prompt = ""
-        for message in conversation:
-            if message["role"] == "user":
-                prompt += f"Human: {message['content']}\n\n"
-            elif message["role"] == "assistant":
-                prompt += f"Assistant: {message['content']}\n\n"
-        formatted_prompts.append(prompt.strip())
-
-    return formatted_prompts
-
-
-def reload_model_weights(llm: LLM, ckpt_path: str):
-    # Access the internal model from vLLM
-    model = llm.llm_engine.model_executor.driver_worker.model_runner.model
-    # Load state dict
-    with safe_open(ckpt_path, framework="pt", device="cpu") as f:
-        # Create a better weight iterator that filters out empty keys and handles prefixes
-        def weights_iterator():
-            for key in f.keys():
-                # Skip empty keys
-                if not key:
-                    continue
-                yield key, f.get_tensor(key)
-
-        # Load weights
-        model.load_weights(weights_iterator())
-
-    # Process weights after loading (important for some models)
-    model_config = llm.llm_engine.model_config
-    device = next(model.parameters()).device
-    _process_weights_after_loading(model, model_config, device)
-
-    return llm
-
-
-def generate_target_length_prompts(config: Config, batch_size: int):
-    len_reward = config.rewards.len_reward
-    if len_reward is None:
-        return [""] * batch_size, [-1] * batch_size
-
-    if len_reward.target_length_sampling == "discrete":
-        indices = torch.randint(low=0, high=len(len_reward.target_lengths), size=(batch_size,), device="cpu")
-        target_lengths = [int(len_reward.target_lengths[i]) for i in indices]
-
-    elif len_reward.target_length_sampling == "range":
-        target_lengths = torch.randint(
-            low=len_reward.min_length, high=len_reward.max_length + 1, size=(batch_size,), device="cpu"
-        ).tolist()
-
-    else:
-        raise ValueError("'length_target_sampling' has to be 'discrete' or 'range'")
-
-    prompt_prefix = " " if len_reward.length_prompt_location == "instruction" else " "
-    max_word = " maximally " if len_reward.reward_type == "clip" else ""
-
-    return [f"{prompt_prefix}Think for{max_word}{target} tokens before giving a response." for target in target_lengths], target_lengths
-
-
 def inference(config: Config):
     # Initialize the logger
     logger.info("Starting inference")
-    logger.info(f"TP={config.tp}, DP={config.dp}, PP={config.pp.world_size}")
 
-    # Initialize prime metrics
+    # Log relevant configuration
+    logger.info(f"Model: {config.model_name}")
+    logger.info(f"Dataset: {config.dataset}")
+    logger.info(f"Parallelism: TP={config.tp}, DP={config.dp}, PP={config.pp.world_size}")
+
+    if config.clean_output_path and config.output_path is not None:
+        logger.debug(f"Cleaning output path {config.output_path}")
+        shutil.rmtree(config.output_path, ignore_errors=True)
+
+    # Initialize metrics
     prime_metric = PrimeMetric(disable=config.prime_log_freq is None, period=config.prime_log_freq)
 
     # Initialize vLLM and get tokenizer
-    logger.info("Initializing vLLM")
+    logger.info(
+        f"Initializing vLLM for {config.model_name} (max_model_len={config.max_model_len}, enforce_eager={config.enforce_eager}, dtype={config.dtype}, quant={config.quant})"
+    )
     llm = LLM(
         model=config.model_name,
         tensor_parallel_size=config.tp,
@@ -135,39 +85,52 @@ def inference(config: Config):
         )
 
     # Load  dataset
-    logger.info(f"Loading dataset {config.dataset}")
     dataset = load_dataset(config.dataset, split="train")
+    logger.info(f"Loaded dataset {config.dataset} with {len(dataset):,} problems")
 
     # Optionally shuffle dataset
-    if envs.NODE_ADDRESS is not None:
+    if envs.GROUP_ID is not None:
         # We dont shuffle here because we shuffle reproducibly in the sampling loop.
-        assert config.seed is None, "Seed is not supported when NODE_ADDRESS is set"
-        assert envs.RANK == 0, "DP is not supported when NODE_ADDRESS is set"
-        node_address_int = int(envs.NODE_ADDRESS, 16)
-        logger.info(f"Seeding with {node_address_int} ({envs.NODE_ADDRESS})")
+        assert config.seed is None, "Seed is not supported when GROUP_ID is set"
+        assert os.environ.get("DP_RANK") is None, "DP is not supported when GROUP_ID is set"
+        node_address_int = int(envs.GROUP_ID, 16)
+        logger.info(f"Seeding with {node_address_int} ({envs.GROUP_ID})")
     else:
         # Seed the dataset with a random number
-        seed = config.seed + envs.RANK if config.seed is not None else None
+        seed = config.seed + int(os.environ.get("DP_RANK", 0)) if config.seed is not None else None
         generator = np.random.default_rng(seed)
-        dataset = load_dataset(config.dataset, split="train").shuffle(generator=generator)
+        logger.info(f"Shuffling dataset with seed {seed}")
+        dataset = dataset.shuffle(generator=generator)
         node_address_int = None
+
+    if config.max_prompt_len:
+        dataset = filter_data_by_prompt_length(dataset, config.max_prompt_len, tokenizer)
+        logger.info(f"✨ Removed long prompts - {len(dataset)} samples remaining")
 
     # Optionally filter dataset
     if config.difficulty_filtering:
+        logger.info(
+            f"Filtering dataset for difficulty in [{config.difficulty_filtering.min_solve_rate}, {config.difficulty_filtering.max_solve_rate}]"
+        )
         dataset = dataset.filter(
             lambda x: x[config.difficulty_filtering.solve_rate_field] >= config.difficulty_filtering.min_solve_rate
             and x[config.difficulty_filtering.solve_rate_field] <= config.difficulty_filtering.max_solve_rate
         )
 
     # Setup TOPLOC
+    num_batch_samples = config.batch_size * config.sampling.n
+    hidden_size = llm.llm_engine.model_executor.driver_worker.model_runner.model.config.hidden_size
     toploc_cache, _ = setup_toploc_cache(
         llm,
         disable=not config.toploc,
-        max_seqs=config.batch_size * config.sampling.n,
-        hidden_size=llm.llm_engine.model_executor.driver_worker.model_runner.model.config.hidden_size,
+        max_seqs=num_batch_samples,
+        hidden_size=hidden_size,
     )
 
+    ckpt_step = 0
+    real_step = 0
     if config.ckpt_start_path is not None:
+        logger.info(f"Resuming from checkpoint {config.ckpt_start_path}")
         path = Path(config.ckpt_start_path)
         path_file = path / "model.safetensors"
         if not path_file.exists():
@@ -176,13 +139,11 @@ def inference(config: Config):
         logger.info(f"Resuming from step {ckpt_step} at {path_file}")
         llm = reload_model_weights(llm, path_file)
         real_step = ckpt_step
-    else:
-        ckpt_step = 0
-        real_step = 0
 
     # This is used by the seeding logic to make sure we dont generate the same samples twice if we do multiple batches for a step
     current_step_batch_counter = 1
     total_problems = 0
+    total_samples = 0
     total_tokens = 0
     max_samples = config.max_samples or len(dataset)
 
@@ -202,20 +163,19 @@ def inference(config: Config):
             else:
                 current_step_batch_counter += 1
 
-        logger.info(
-            f"real_step: {real_step}, ckpt_step: {ckpt_step}, real_step - ckpt_step: {real_step - ckpt_step}, config.async_level: {config.async_level}"
-        )
+        logger.info(f"Inference step {real_step} (Checkpoint step: {ckpt_step})")
         if config.rollout_path is not None and real_step - ckpt_step > config.async_level:
+            logger.info(f"Required to reload model weights for step {ckpt_step} from {config.rollout_path}")
             ckpt_step = real_step - config.async_level
             attempt_count = 0
             while True:
                 stable_file = Path(config.rollout_path) / f"step_{ckpt_step}/stable"
                 if stable_file.exists():
-                    logger.info(f"Reloading model weights from {config.rollout_path} ckpt {ckpt_step}")
+                    logger.info(f"Reloading model weights for step {ckpt_step} from {stable_file}")
                     llm = reload_model_weights(llm, Path(config.rollout_path) / f"step_{ckpt_step}/model.safetensors")
                     total_problems = 0
                     total_tokens = 0
-                    logger.info(f"Reloaded model weights from {config.rollout_path} ckpt {ckpt_step}")
+                    logger.info(f"Reloaded model weights for step {ckpt_step} from {stable_file}")
                     break
                 if attempt_count % 30 == 0:
                     logger.info(f"No stable file found at {stable_file}, waiting for new checkpoint")
@@ -230,13 +190,16 @@ def inference(config: Config):
             # We reseed the generator here to make the sampling reproducible at each step.
             # This would work even if the node restarts and resumes from the current step.
             generator = np.random.default_rng(node_address_int * current_step_batch_counter + real_step)
-            indexes = generator.integers(0, len(dataset), config.batch_size)
-            batch = dataset.select(indexes)
+            indices = generator.integers(0, len(dataset), config.batch_size)
         else:
-            batch = dataset.select(range(i, min(i + config.batch_size, len(dataset))))
+            indices = list(range(i, min(i + config.batch_size, len(dataset))))
+
+        logger.debug(f"Sampling batch with indices [{' '.join(map(str, indices[:3]))}...{' '.join(map(str, indices[-3:]))}]")
+        batch = dataset.select(indices)
+
         messages = [[{"role": "user", "content": item["prompt"]}, {"role": "assistant", "content": "<think>\n"}] for item in batch]
 
-        length_prompt_additions, target_lengths = generate_target_length_prompts(config, len(batch))
+        length_prompt_additions, target_lengths = generate_target_length_prompts(config.len_reward, len(batch))
         # Assume verification_info is stored as a JSON string in the dataset.
         verification_infos = [json.loads(item["verification_info"]) for item in batch]
         for target_length, verification_info in zip(target_lengths, verification_infos):
@@ -277,7 +240,7 @@ def inference(config: Config):
         request_outputs = llm.generate(prompts, sampling_params, use_tqdm=False)
         end_time = time.time()
 
-        # Dropping like this isnt ideal. But in practice, we shouldnt have any prompts that are too long.
+        # Dropping like this isn't ideal. But in practice, we shouldn't have any prompts that are too long.
         request_outputs = [req for req in request_outputs if len(req.outputs[0].token_ids) > 0]
         if len(request_outputs) != len(prompts):
             logger.warning(f"{len(prompts) - len(request_outputs)} prompts were filtered out because they were too long")
@@ -286,19 +249,23 @@ def inference(config: Config):
         # We call here to give time for the proofs to be generated non-blocking in the background.
         toploc_cache.maybe_generate_proofs_in_background(force_generate=True)
 
-        # Calculate tokens and throughput
+        # Calculate batch problems, samples and tokens
+        batch_problems = len(batch)
+        batch_samples = sum(len(req.outputs) for req in request_outputs)
         batch_input_tokens = sum(len(req.prompt_token_ids) for req in request_outputs)
         batch_output_tokens = sum(sum(len(output.token_ids) for output in req.outputs) for req in request_outputs)
-        batch_total_tokens = batch_input_tokens + batch_output_tokens
-        total_tokens += batch_total_tokens
+        batch_tokens = batch_input_tokens + batch_output_tokens
+        # Calculate overall problems, samples and tokens
+        total_tokens += batch_tokens
+        total_problems += batch_problems
+        total_samples += batch_samples
+        logger.info(f"Generated {batch_samples} samples for {batch_problems} problems for step {real_step} in {end_time - start_time:.2f}s")
 
-        avg_seq_length = batch_total_tokens / (len(request_outputs) * config.sampling.n) if request_outputs else 0
-
-        elapsed_time = end_time - start_time
-        tokens_per_second = batch_total_tokens / elapsed_time if elapsed_time > 0 else 0
-
+        # Compute batch throughput and average sequence length
+        batch_throughput = batch_tokens / (end_time - start_time)
+        avg_sequence_length = batch_tokens / num_batch_samples
         logger.info(
-            f"Batch throughput: {tokens_per_second:.2f} tok/sec ({batch_total_tokens} tokens in {elapsed_time:.2f}s, avg seq len: {avg_seq_length:.1f})"
+            f"Batch throughput: {batch_throughput:.2f} tok/sec ({batch_tokens} tokens in {end_time - start_time:.2f}s, avg seq len: {avg_sequence_length:.1f})"
         )
 
         # Compute proofs
@@ -312,7 +279,7 @@ def inference(config: Config):
         # Compute rewards and advantages
         start = time.time()
         request_rewards = compute_vllm_rewards(request_outputs, verification_infos, task_types, config.rewards)
-        logger.info(f"Computed rewards and advantages in in {time.time() - start:.2f}s")
+        logger.info(f"Computed rewards and advantages in {time.time() - start:.2f}s")
 
         table = get_parquet_table(
             request_outputs,
@@ -326,20 +293,21 @@ def inference(config: Config):
         os.makedirs(step_path, exist_ok=True)
         pq_save_path = f"{step_path}/{uuid.uuid4()}.parquet"
         pq.write_table(table, pq_save_path)
+        logger.info(f"Saved batch outputs to {pq_save_path}")
+
         file_sha = sha256sum(pq_save_path)
         prime_metric.log_prime({"file_sha": file_sha, "file_name": pq_save_path})
-        logger.info(f"✨ Saved {len(proofs)} samples to {pq_save_path} with sha {file_sha or 'NA'}")
 
-        total_problems += len(prompts)
         metric = {"dashbord-progress/total": total_problems, f"dashbord-progress/{config.dataset}": total_tokens}
         prime_metric.log_prime(metric)
 
-        logger.info(f"Generated {total_problems} problems for step {real_step}")
         real_step += 1
 
         if config.total_step is not None and real_step > config.total_step:
             logger.info(f"Reached total step {config.total_step}, stopping inference")
             break
+
+    logger.info(f"Inference finished! Generated {total_samples} samples for {total_problems} problems")
 
     # Manually destroy vLLM process group to avoid warnings
     dist.destroy_process_group()
@@ -347,16 +315,16 @@ def inference(config: Config):
 
 def main(config: Config) -> list[mp.Process]:
     processes = []
-    from zeroband.inference import envs as inference_envs
+    import zeroband.inference.envs as envs
 
     if config.dp > 1:
         if config.tp == "auto":
             assert torch.cuda.device_count() % config.dp == 0, "Number of GPUs must be divisible by DP"
             config.tp = torch.cuda.device_count() // config.dp
-        gpu_ids = inference_envs.CUDA_VISIBLE_DEVICES
+        gpu_ids = envs.CUDA_VISIBLE_DEVICES
         gpu_ids_per_rank = [gpu_ids[i : i + config.tp] for i in range(0, len(gpu_ids), config.tp)]
         for rank, gpu_ids in enumerate(gpu_ids_per_rank):
-            envs = {"CUDA_VISIBLE_DEVICES": ",".join(map(str, gpu_ids)), "RANK": str(rank), "LOCAL_RANK": str(rank)}
+            envs = {"CUDA_VISIBLE_DEVICES": ",".join(map(str, gpu_ids)), "DP_RANK": str(rank)}
             process = mp.Process(target=EnvWrapper(inference, envs), args=(config,))
             processes.append(process)
     else:
@@ -392,7 +360,7 @@ if __name__ == "__main__":
             inference_envs.SHARDCAST_SERVERS,
             config.rollout_path,
             config.async_level + 1,
-            # TODO: maybe +1 because we most likely wont download the current step in time?
+            # TODO: maybe +1 because we most likely won't download the current step in time?
             # We could deadlock though.
             max(current_step - config.async_level, 1),
         )
