@@ -1,45 +1,41 @@
+import logging
 import os
-from pathlib import Path
 import shutil
 import time
+from pathlib import Path
 from typing import TYPE_CHECKING
 
+import shardcast
 import torch
 import torch.distributed as dist
-from torch.distributed._composable.fsdp import fully_shard, MixedPrecisionPolicy  # type: ignore
 import torch.distributed.tensor
 import wandb
-import shardcast
+from jaxtyping import Float
+from liger_kernel.transformers import apply_liger_kernel_to_qwen2
+from pydantic_config import parse_argv
+from torch._guards import log as torch_log
+from torch.distributed._composable.fsdp import MixedPrecisionPolicy, fully_shard  # type: ignore
 
-from zeroband.training.config import Config
-from zeroband.utils.models import ModelType, get_model_and_tokenizer
 from zeroband.training import envs
 from zeroband.training.checkpoint import TrainingProgress, load_checkpoint_fsdp_state, save_checkpoint_fsdp_state, save_ckpt_for_rollout
+from zeroband.training.config import Config
 from zeroband.training.data import BatchOutput, DatasetOutput, get_dataloader, packed_batch
-from zeroband.training.loss import grpo_loss, kl_penalty, selective_log_softmax, entropy_loss
+from zeroband.training.loss import entropy_loss, grpo_loss, kl_penalty, selective_log_softmax
 from zeroband.training.lr_scheduler import get_scheduler
 from zeroband.training.utils import (
+    MetricsAverager,
     PerfCounter,
     apply_ac_ckpt,
-    MetricsAverager,
+    log_prompt_response_samples,
+    log_to_wandb,
     offload_model_to_cpu,
     reshard_module,
     wake_up_model_from_cpu,
 )
-
-from zeroband.utils.logger import get_logger
-
-from pydantic_config import parse_argv
-from jaxtyping import Float
-
-from zeroband.utils.world_info import WorldInfo, get_world_info
-
-
-from liger_kernel.transformers import apply_liger_kernel_to_qwen2
-from torch._guards import log as torch_log
-import logging
-
+from zeroband.training.world_info import WorldInfo, get_world_info
 from zeroband.utils.http_monitor import HttpMonitor
+from zeroband.utils.logger import get_logger
+from zeroband.utils.models import ModelType, get_model_and_tokenizer
 
 
 def get_local_batch_size(batch_size: int, micro_bs: int, data_workers: int, world_info: WorldInfo) -> int:
@@ -93,6 +89,11 @@ def train(config: Config):
 
     logger = get_logger("TRAIN")
     world_info = get_world_info()
+    wandb_sample_history = None
+
+    if config.ckpt.clean_rollout_path and config.ckpt.rollout_path is not None:
+        logger.info(f"Cleaning rollout path {config.ckpt.rollout_path}")
+        shutil.rmtree(config.ckpt.rollout_path, ignore_errors=True)
 
     logger.info(f"start training on {world_info.world_size} rank(s)")
 
@@ -146,13 +147,11 @@ def train(config: Config):
         num_training_steps=config.optim.total_steps,
     )
 
-    total_problems = (
-        config.start_total_problems if config.start_total_problems is not None else config.optim.batch_size * config.optim.total_steps
-    )
-    training_progress = TrainingProgress(total_tokens=0, step=config.start_step, total_problems=total_problems)
+    total_samples = config.start_total_samples if config.start_total_samples is not None else 0
+    training_progress = TrainingProgress(total_tokens=0, step=config.start_step, total_samples=total_samples)
 
     if world_info.rank == 0 and config.wandb:
-        wandb.init(project=config.project, config=config.model_dump())
+        wandb.init(project=config.project, config=config.model_dump(), dir="wandb_logs", name=config.wandb_run_name)
 
     if envs.PRIME_API_BASE_URL is not None:
         monitor = HttpMonitor()
@@ -254,6 +253,8 @@ def train(config: Config):
             logger.info(f"Time to compute logprobs: {time_logprob:.2f} seconds")
 
         logger.debug("start training rollout")
+
+        # In the training loop
         for rollout_step in range(config.optim.step_per_rollout):
             logger.debug(f"training rollout step {rollout_step} / {config.optim.step_per_rollout}")
             metric_averager = MetricsAverager()
@@ -265,6 +266,19 @@ def train(config: Config):
             data_per_rollout = next(logprobs_aware_iterator)
             num_grad_acc_steps = len(data_per_rollout)
 
+            # Collect samples for WandB logging - do this ONCE per step
+            if world_info.rank == 0 and config.wandb:
+                # Use the first batch for logging (could be configurable if needed)
+                batch = data_per_rollout[0]
+
+                # Log the samples to WandB with history management
+                try:
+                    # Pass and update the sample history
+                    wandb_sample_history = log_prompt_response_samples(tokenizer, batch, training_progress.step, wandb_sample_history)
+                except Exception as e:
+                    logger.warning(f"Error logging samples to WandB: {e}")
+
+            # Now here's the complete grad_acc_step loop WITHOUT the WandB logging inside it:
             for grad_acc_step in range(num_grad_acc_steps):
                 logger.debug(f"training grad_acc_step {grad_acc_step} / {num_grad_acc_steps}")
                 batch = data_per_rollout[grad_acc_step]
@@ -285,14 +299,6 @@ def train(config: Config):
                     metric_averager.update("length_penalties", length_penalties)
                 for target_lengths in batch["target_lengths"]:
                     metric_averager.update("target_lengths", target_lengths)
-
-                ## per micro batch metrics
-
-                metric_averager.update("batch_reward", batch["rewards"].float().mean())
-                metric_averager.update("batch_task_reward", batch["task_rewards"].float().mean())
-                metric_averager.update("batch_seq_lens", batch["seq_lens"].float().mean())
-                metric_averager.update("batch_length_penalties", batch["length_penalties"].float().mean())
-                metric_averager.update("batch_target_lengths", batch["target_lengths"].float().mean())
 
                 # Forward
                 logits: Float[torch.Tensor, "batch seq vocab"] = model(
@@ -331,6 +337,8 @@ def train(config: Config):
                 loss = loss / num_grad_acc_steps
 
                 inputs_ids_shape = input_ids.shape
+
+                # Now we can delete the batch data
                 del batch, logits, input_ids, advantages, loss_mask, original_logprobs
 
                 # Backward
@@ -364,7 +372,7 @@ def train(config: Config):
             new_tokens = world_info.world_size * token_per_gpu
             perf_counter.count_tokens(new_tokens)
             training_progress.total_tokens += new_tokens
-            training_progress.total_problems += config.optim.batch_size
+            training_progress.total_samples += config.optim.batch_size
 
             padding_proportion = (config.data.seq_length - metric_averager["seq_lens"].item() - 1) / config.data.seq_length
 
@@ -378,7 +386,7 @@ def train(config: Config):
                 "grad_norm": grad_norm.item(),
                 "padding_proportion": padding_proportion,
                 "grad_acc_steps": num_grad_acc_steps,
-                "total_problems": training_progress.total_problems,
+                "total_samples": training_progress.total_samples,
             }
 
             for key, value in metric_averager.items():
@@ -406,7 +414,7 @@ def train(config: Config):
 
             if world_info.rank == 0:
                 if config.wandb:
-                    wandb.log(metrics)
+                    log_to_wandb(metrics)
                 if envs.PRIME_API_BASE_URL is not None:
                     monitor.log(metrics)
 
@@ -447,11 +455,15 @@ def train(config: Config):
         time_rollout_step = time.time() - time_start
         logger.info(f"Finished rollout {rollout_step} step {training_progress.step}")
         if world_info.rank == 0 and config.wandb:
-            new_metrics = {"rollout_step": rollout_step, "step": training_progress.step, "time_rollout_step": time_rollout_step}
-            new_metrics["time_logprob"] = time_logprob
-            new_metrics["time_data_loading"] = total_time_data_loading
-            new_metrics["time_packing"] = total_time_packing
-            wandb.log(new_metrics)
+            new_metrics = {
+                "rollout_step": rollout_step,
+                "step": training_progress.step,
+                "time_rollout_step": time_rollout_step,
+                "time_logprob": time_logprob,
+                "time_data_loading": total_time_data_loading,
+                "time_packing": total_time_packing,
+            }
+            log_to_wandb(new_metrics)
 
         if training_progress.step >= config.optim.total_steps:
             break
