@@ -1,10 +1,10 @@
-import pytest
 from multiprocessing import Process, Queue
+
+import pytest
+from prime_iroh import Node
 
 import zeroband.utils.envs as envs  # noqa
 from zeroband.inference.pipeline import setup_comm
-
-from prime_iroh import Node
 
 # Pre-computed node IDs for different seeds (our team's favorite numbers)
 IROH_NODE_ID_MAP = {
@@ -42,6 +42,7 @@ def _setup_comm(rank: int, world_size: int, error_queue: Queue):
 
 
 @pytest.mark.parametrize("world_size", [1, 2, 4, 8])
+@pytest.mark.slow
 def test_setup_comm(world_size: int):
     # Test that setup_comm raises an error for 1 stage
     if world_size == 1:
