@@ -3,22 +3,9 @@ import subprocess
 
 import pytest
 
-from zeroband.utils.world_info import get_world_info, reset_world_info
+from zeroband.training.world_info import get_world_info
 
 ENV_VARS = ["RANK", "WORLD_SIZE", "LOCAL_RANK", "LOCAL_WORLD_SIZE"]
-
-
-def reset_env_vars():
-    for var in ENV_VARS:
-        if var in os.environ:
-            del os.environ[var]
-
-
-@pytest.fixture(autouse=True)
-def setup():
-    reset_env_vars()
-    reset_world_info()
-    yield
 
 
 def test_init_with_default_args():
@@ -57,7 +44,7 @@ def test_init_with_valid_env_vars(local_world_size: int, world_size: int):
 
 @pytest.mark.parametrize("local_world_size", [1, 2])
 def test_init_with_torchrun(local_world_size: int):
-    path = "src/zeroband/utils/world_info.py"
+    path = "src/zeroband/training/world_info.py"
     assert os.path.exists(path)
     cmd = ["torchrun", f"--nproc_per_node={local_world_size}", path]
     process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
